# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
paths:
...
  /api/fitness/analyze:
    post:
      operationId: analyzeFitness
      summary: Run harness fluency analysis and return the additive harnessability baseline for one or more profiles
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Fitness fluency result payload with additive harnessability baseline data
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid fitness analysis context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Fitness analysis failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/a2ui/dashboard:
    get:
      operationId: getA2UIDashboard
      summary: Get A2UI v0.10 dashboard data
      description: |
        Returns A2UI v0.10 messages for the workspace dashboard.
        Aggregates sessions, tasks, background tasks, traces, and custom surfaces.
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: string
            default: "default"
          description: Workspace ID (defaults to 'default')
      responses:
        "200":
          description: A2UI dashboard data
          content:
            application/json:
              schema:
                type: object
                required: [data, kind]
                properties:
                  data:
                    type: array
                    description: Array of A2UI v0.10 messages
                    items:
                      type: object
                  kind:
                    type: string
                    const: "data"
                  metadata:
                    type: object
                    properties:
                      mimeType:
                        type: string
                        example: "application/json+a2ui"
        "500":
          description: Failed to generate A2UI dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: postA2UIDashboardMessage
      summary: Add custom A2UI messages to the dashboard
      description: |
        Accepts custom A2UI messages and stores them for the workspace.
        Messages are appended to the dashboard data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                workspaceId:
                  type: string
                  default: "default"
                  description: Workspace ID to add messages to
                messages:
                  type: array
                  description: Custom A2UI v0.10 messages
                  items:
                    type: object
      responses:
        "200":
          description: Messages added successfully
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Failed to store messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── AG-UI Protocol ──
...